+++ addOns/dev/CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
  - Login where the session token is base64 encoded when passed to the browser.
  - HTML page with input elements added with increasing delays.
  - A mock MS online login.
+ - Login using HTTP basic authentication.
 
 ### Changed
 - Update minimum ZAP version to 2.17.0.